+++ CHANGELOG.md
@@ -6,8 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 ### Changed
-- Implemented Wellness category pages deep linking [#4504](https://github.com/rokwire/illinois-app/issues/4504).
 - Updated wellness resources content [#4504](https://github.com/rokwire/illinois-app/issues/4504).
+### Added
+- Added Wellness category pages deep linking [#4504](https://github.com/rokwire/illinois-app/issues/4504).
+- Added the "Message this user" in Directory Account card [#4555](https://github.com/rokwire/illinois-app/issues/4555).
 
 ## [6.1.31] - 2024-12-20
 ### Changed